python/requirements.txt
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
openai>=1.0
tqdm>=4.0
python-dotenv>=1.0
pytest>=8.0
pdfplumber>=0.11
python-pptx>=1.0
openpyxl>=3.1
python-docx>=1.0
olefile>=0.47
